Settings: move airdrop to testnet/airdrop
